backend/tests/requirements.txt
django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
pyyaml==5.4.1
django-cors-headers==3.7.0